requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
apscheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
dateparser==1.2.0
transformers==4.35.2
torch>=2.0.0

